1.6.0 beta 2
1. Add debug gizmo support for seeing the layout of 3CC.
2. BasePoolStateManager has been merged with PoolStateManager.
3. PoolStateManager is now broken out into a set of partial files which are far easier to maintain and read.
4. Added a nice custom Inspector for PoolStateManager. Thanks Vowgan!
5. Cue ball jumping off the table is now much more cleanly handled (mostly inside CueBallOffTableController), fixing a bug.
6. You can now always force-end a game if you're instance owner or instance master. Sadly, there's no Groups support in Udon so we can't extend this to group moderators.
7. The game mode you're playing is now an enum, which makes it far, far, far cleaner to maintain going forward.
8. The desktop UI is now actually on the screen.
9, Three Cushion Carom is now selectable in all three menus.
10. The M.O.O.N UI has been somewhat modified to allow for Three Cushion Carom; this was unavoidable as it was hard-coded to assume only four game modes.
11. Three Cushion Carom is now actually the right game mode, and not a cursed game with the wrong number of balls and the wrong rules. The game now uses three balls, handles using the opponent's cue ball an object ball properly, handles scoring on breaks, correctly positions balls on breaks and table resets, etc.
12. Other assorted bug fixes.

KNOWN ISSUES
1. The intro animation is broken. This is fine as it's being deprecated anyway.
2. This requires a lot of testing outside the Unity editor, especially in networked environments.
3. The Three Cushion Carom icon on the Esnya & Akalink menus is temporary.
4. iOS users, if you exist, may notice the camera drags too slowly. This will be fixed as soon as there's proof of anyone playing VRC on iOS.
