+++ Packages/com.vrcbilliards.vrcbce/InternalComponents/Desktop UI.prefab
@@ -187,8 +187,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 1, y: 1}
-  m_AnchoredPosition: {x: 1.2276611, y: 38.929962}
-  m_SizeDelta: {x: -20, y: 6.1398926}
+  m_AnchoredPosition: {x: 1.2276611, y: 88.300125}
+  m_SizeDelta: {x: -20, y: -263.39975}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &3318120536040607545
 CanvasRenderer:
@@ -881,7 +881,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0.5}
   m_AnchorMax: {x: 0, y: 0.5}
-  m_AnchoredPosition: {x: 250, y: 0}
+  m_AnchoredPosition: {x: 100, y: -25}
   m_SizeDelta: {x: 200, y: 1040}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!1 &3200909986064807058
@@ -2761,7 +2761,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!4 &8405257728537727517
 Transform:
   m_ObjectHideFlags: 0
@@ -3051,7 +3051,7 @@ Camera:
     width: 1
     height: 1
   near clip plane: 0.01
-  far clip plane: 10
+  far clip plane: 1000
   field of view: 60
   orthographic: 0
   orthographic size: 1.25
@@ -3143,7 +3143,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   serializationData:
-    SerializedFormat: 0
+    SerializedFormat: 2
     SerializedBytes: 
     ReferencedUnityObjects: []
     SerializedBytesString:
