+++ Packages/com.vrcbilliards.vrcbce/InternalComponents/Screenspace UI.prefab
@@ -34,8 +34,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
-  m_AnchoredPosition: {x: 271, y: 60}
-  m_SizeDelta: {x: 100, y: 100}
+  m_AnchoredPosition: {x: 200.15332, y: 41.937897}
+  m_SizeDelta: {x: 73.2553, y: 67.1598}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &3398507934973257016
 CanvasRenderer:
@@ -109,8 +109,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
-  m_AnchoredPosition: {x: 313, y: 49}
-  m_SizeDelta: {x: 100, y: 100}
+  m_AnchoredPosition: {x: 230.92053, y: 34.550293}
+  m_SizeDelta: {x: 73.2553, y: 67.1598}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &6848012847749805731
 CanvasRenderer:
@@ -184,8 +184,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
-  m_AnchoredPosition: {x: 135, y: 51}
-  m_SizeDelta: {x: 200, y: 50}
+  m_AnchoredPosition: {x: 100.52612, y: 35.893494}
+  m_SizeDelta: {x: 146.5106, y: 33.5799}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &542511724924032877
 CanvasRenderer:
@@ -243,7 +243,7 @@ MonoBehaviour:
   m_faceColor:
     serializedVersion: 2
     rgba: 4294967295
-  m_fontSize: 28.35
+  m_fontSize: 20.75
   m_fontSizeBase: 36
   m_fontWeight: 400
   m_enableAutoSizing: 1
@@ -346,7 +346,7 @@ Canvas:
   m_OverrideSorting: 0
   m_OverridePixelPerfect: 0
   m_SortingBucketNormalizedSize: 0
-  m_VertexColorAlwaysGammaSpace: 0
+  m_VertexColorAlwaysGammaSpace: 1
   m_AdditionalShaderChannelsFlag: 25
   m_UpdateRectTransformForStandalone: 0
   m_SortingLayerID: 0
@@ -482,8 +482,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
-  m_AnchoredPosition: {x: 177.60669, y: 55}
-  m_SizeDelta: {x: 343.0135, y: 100}
+  m_AnchoredPosition: {x: 131.7378, y: 38.579895}
+  m_SizeDelta: {x: 251.2756, y: 67.1598}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &6927788071662898852
 CanvasRenderer:
