+++ Packages/com.vrcbilliards.vrcbce/Materials/CueGrip.mat
@@ -86,7 +86,7 @@ Material:
     - _UVSec: 0
     - _ZWrite: 1
     m_Colors:
-    - _Color: {r: 0.34, g: 0.34, b: 0.34, a: 1}
+    - _Color: {r: 0, g: 0.5, b: 1.1, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
     - _TintColor: {r: 0, g: 0.46255493, b: 1, a: 0.07058824}
   m_BuildTextureStacks: []
