+++ Packages/com.vrcbilliards.vrcbce/Runtime/Scripts/PoolPositioner.asset
@@ -62,7 +62,7 @@ MonoBehaviour:
       Data: 3|System.RuntimeType, mscorlib
     - Name: 
       Entry: 1
-      Data: VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts.BasePoolStateManager,
+      Data: VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts.PoolStateManager,
         VRCBilliards.VRCBCE.Runtime
     - Name: 
       Entry: 8
