+++ Packages/com.vrcbilliards.vrcbce/Runtime/Scripts/PoolStateManager.Base.cs
@@ -1,19 +1,11 @@
 ﻿using System;
-using System.Runtime.Remoting.Messaging;
 using UdonSharp;
 using UnityEngine;
 using UnityEngine.Animations;
 using UnityEngine.Rendering;
-using UnityEngine.Serialization;
 using VRC.SDK3.Components;
-using VRC.SDK3.Data;
 using VRC.SDKBase;
-using VRC.Udon;
 using VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts.Components;
-
-// ReSharper disable SpecifyACultureInStringConversionExplicitly
-// ReSharper disable CheckNamespace
-// ReSharper disable InconsistentNaming
 
 namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
 {
@@ -28,13 +20,20 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         PlayerLeft
     }
 
+    public enum GameMode {
+        EightBall,
+        NineBall,
+        KoreanCarom,
+        JapaneseCarom,
+        ThreeCushionCarom
+    }
+
     /// <summary>
     /// This is the base logic that governs the pool table, devoid of almost all physics code. This code is quite
     /// messy; it includes all setup, teardown, game state and replication logic, and works with all other components
     /// to allow the player to play pool.
     /// </summary>
-    [UdonBehaviourSyncMode(BehaviourSyncMode.Manual)]
-    public abstract class BasePoolStateManager : DebuggableUdon
+    public partial class PoolStateManager
     {
         #region Dependencies
 
@@ -48,24 +47,24 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
 
         #region Networked Variables
 
-        [UdonSynced] protected bool turnIsRunning;
+        [UdonSynced] private bool turnIsRunning;
         [UdonSynced] private int timerSecondsPerShot;
         [UdonSynced] private bool isPlayerAllowedToPlay;
         [UdonSynced] private bool guideLineEnabled = true;
 
-        [UdonSynced] protected bool[] ballsArePocketed;
+        [UdonSynced] private bool[] ballsArePocketed;
         private bool[] oldBallsArePocketed;
 
-        [UdonSynced] protected bool isTeam2Turn;
+        [UdonSynced] private bool isTeam2Turn;
         private bool oldIsTeam2Turn;
 
         [UdonSynced] private bool isFoul;
         [UdonSynced] private bool isGameOver;
 
-        [UdonSynced] protected bool isOpen = true;
+        [UdonSynced] private bool isOpen = true;
         private bool oldOpen;
 
-        [UdonSynced] protected bool isTeam2Blue;
+        [UdonSynced] private bool isTeam2Blue;
 
         [UdonSynced] private bool isGameInMenus = true;
         private bool oldIsGameInMenus;
@@ -79,17 +78,18 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
 
         [UdonSynced] private uint turnID;
         [UdonSynced] private bool isRepositioningCueBall;
+        
+        /// <summary>
+        /// Is the game currently in the first shot?
+        /// </summary>
+        [UdonSynced] private bool _isGameBreak;
 
         [UdonSynced] private int[] scores = new int[2];
-        [UdonSynced] protected Vector3[] currentBallPositions = new Vector3[NUMBER_OF_SIMULATED_BALLS];
-        [UdonSynced] protected Vector3[] currentBallVelocities = new Vector3[NUMBER_OF_SIMULATED_BALLS];
-        [UdonSynced] protected Vector3[] currentAngularVelocities = new Vector3[NUMBER_OF_SIMULATED_BALLS];
+        [UdonSynced] private Vector3[] currentBallPositions = new Vector3[NUMBER_OF_SIMULATED_BALLS];
+        [UdonSynced] private Vector3[] currentBallVelocities = new Vector3[NUMBER_OF_SIMULATED_BALLS];
+        [UdonSynced] private Vector3[] currentAngularVelocities = new Vector3[NUMBER_OF_SIMULATED_BALLS];
 
-        // TODO: These gameMode vars are sloppy and are a good candidate for a refactor into one value, and really
-        // this can just be a string.
-        [UdonSynced] protected uint gameMode; // 0 = 8ball, 1 = 9ball, 2 = Carom
-        [UdonSynced] private bool isKorean; // False = Japanese 4-Ball Carom, True = Korean 4-Ball Carom
-        [UdonSynced] private bool isThreeCushionCarom;
+        [UdonSynced] private GameMode gameMode;
         [UdonSynced] private int player1ID;
         [UdonSynced] private int player2ID;
         [UdonSynced] private int player3ID;
@@ -133,14 +133,17 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         #endregion
 
         // The number of balls we simulate - this const allows us to increase the number we support in the future.
-        protected const int NUMBER_OF_SIMULATED_BALLS = 16;
+        private const int NUMBER_OF_SIMULATED_BALLS = 16;
 
         // A small fraction designed to slightly move balls around when placed, which helps with making the table
         // less deterministic.
-        protected const float RANDOMIZE_F = 0.0001f;
+        private const float RANDOMIZE_F = 0.0001f;
 
-        // These four consts dictate some placement of balls when placed.
+        // These four vars dictate some placement of balls when placed.
+        
+        // Half-way between the centre spot and the cushion on the X axis.
         public float SPOT_POSITION_X = 0.5334f;
+        // Three quarters of the way between the centre spot and the cushion on the X axis.
         public float SPOT_CAROM_X = 0.8001f;
         public float BALL_PL_X = 0.03f;
         public float BALL_PL_Y = 0.05196152422f;
@@ -155,22 +158,21 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
 
         #endregion
 
-        /*[Header("Options")]*/ [Tooltip("Use fake shadows? They may clash with your world's lighting.")]
+        [Tooltip("Use fake shadows? They may clash with your world's lighting.")]
         public bool fakeBallShadows = true;
 
         [Tooltip("Does the table model for this table have rails that guide the ball when the ball sinks?")]
         public bool tableModelHasRails;
 
-        /*[Header("Important Objects")]*/ public Transform sunkBallsPositionRoot;
+        public Transform sunkBallsPositionRoot;
         public GameObject shadows;
         public ParticleSystem plusOneParticleSystem;
-
         public ParticleSystem minusOneParticleSystem;
 
         // Where's the surface of the table?
         public Transform tableSurface;
 
-        /*[Header("Shader Information")]*/ public string uniformTableColour = "_EmissionColor";
+        public string uniformTableColour = "_EmissionColor";
         public string uniformMarkerColour = "_Color";
         public string uniformCueColour = "_EmissionColor";
         public string uniformBallColour = "_BallColour";
@@ -182,7 +184,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         [Range(0f, 5f)]
         public float introAnimationLength = 2.0f;
 
-        /*[Header("Table Colours")]*/ [ColorUsage(true, true)]
+        [ColorUsage(true, true)]
         public Color tableBlue = new Color(0.0f, 0.5f, 1.5f, 1.0f);
 
         [ColorUsage(true, true)] public Color tableOrange = new Color(1.5f, 0.5f, 0.0f, 1.0f);
@@ -201,13 +203,13 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
 
         [ColorUsage(true, true)] private Color[] ballColors = new Color[NUMBER_OF_SIMULATED_BALLS];
 
-        /*[Header("Colour Options")]*/ public bool ballCustomColours;
+        public bool ballCustomColours;
         public ColorPicker blueTeamSliders;
         public ColorPicker orangeTeamSliders;
 
         private float shaderToggleFloat = 0;
 
-        /*[Header("Cues")]*/ protected GameObject cueTip;
+        private GameObject cueTip;
         public GameObject[] cueTips;
         public MeshRenderer[] cueRenderObjs;
         private Material[] cueMaterials = new Material[2];
@@ -224,12 +226,10 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             "\nIf the order of the balls is incorrect, gameplay will not proceed correctly."
         )]
         public Transform[] ballTransforms;
-
         private Rigidbody[] ballRigidbodies;
 
         [Tooltip("The shadow object for each ball")]
         public PositionConstraint[] ballShadowPosConstraints;
-
         private Transform[] ballShadowPosConstraintTransforms;
         public ShotGuideController guideline;
         public GameObject devhit;
@@ -239,7 +239,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         public GameObject marker9ball;
         public GameObject pocketBlockers;
 
-        /*[Header("Materials")]*/ public MeshRenderer[] ballRenderers;
+        public MeshRenderer[] ballRenderers;
 
         public MeshRenderer tableRenderer;
         private Material[] tableMaterials;
@@ -248,7 +248,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
 
         public Material[] cueGrips;
 
-        /*[Header("Audio")]*/ public GameObject audioSourcePoolContainer;
+        public GameObject audioSourcePoolContainer;
         public AudioSource cueTipSrc;
         public AudioClip introSfx;
         public AudioClip sinkSfx;
@@ -273,9 +273,9 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         public AudioClip pointMadeSfx;
         public AudioClip hitBallSfx;
 
-        /*[Header("Reflection Probes")]*/ public ReflectionProbe tableReflection;
+        public ReflectionProbe tableReflection;
 
-        /*[Header("Meshes")]*/ public Mesh[] cueballMeshes;
+        public Mesh[] cueballMeshes;
         public Mesh nineBall;
 
         /// <summary>
@@ -285,11 +285,10 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
 
         private int localPlayerID = -1;
 
-        /*[Header("Desktop Stuff")]*/ public GameObject desktopHitPosition;
+        public GameObject desktopHitPosition;
 
         public GameObject desktopBase;
-
-        //public GameObject desktopQuad;
+        
         public GameObject[] desktopCueParents;
         public UnityEngine.UI.Image tiltAmount;
 
@@ -300,19 +299,16 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
 
         private Transform[] ballPoolTransforms;
         private AudioSource mainSrc;
-        private UdonBehaviour udonChips;
 
         /// <summary>
         /// We are waiting for our local simulation to finish, before we unpack data
         /// </summary>
-        protected bool isUpdateLocked;
+        private bool isUpdateLocked;
 
         /// <summary>
         /// The first ball to be hit by cue ball
         /// </summary>
         private int firstHitBallThisTurn;
-
-        // TODO: This would be a good option.
 
         [Range(10, 50),
          Tooltip("How many points do your players need to win a carom game? This is the same for all carom variants.")]
@@ -320,12 +316,12 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
 
         private int secondBallHitThisTurn;
         private int thirdBallHitThisTurn;
-        protected int cushionsHitThisTurn;
+        private int cushionsHitThisTurn;
 
         /// <summary>
         /// If the simulation was initiated by us, only set from update
         /// </summary>
-        protected bool isSimulatedByUs;
+        private bool isSimulatedByUs;
 
         /// <summary>
         /// Ball dropper timer
@@ -336,16 +332,12 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         private bool isTimerRunning;
         private bool isMadePoint;
         private bool isMadeFoul;
-
-        protected bool is8Ball;
-        protected bool isNineBall;
-        protected bool isFourBall;
+        private bool IsCarom => gameMode == GameMode.KoreanCarom || gameMode == GameMode.JapaneseCarom || gameMode == GameMode.ThreeCushionCarom;
 
         /// <summary>
         /// Game should run in practice mode
         /// </summary>
         private bool isGameModePractice;
-
         private bool isInDesktopTopDownView;
 
         /// <summary>
@@ -397,15 +389,15 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         /// </summary>
         private Vector3 localSpacePositionOfCueTip;
 
-        protected Vector3 localSpacePositionOfCueTipLastFrame;
-        protected Vector3 cueLocalForwardDirection;
-        protected Vector3 cueArmedShotDirection;
+        private Vector3 localSpacePositionOfCueTipLastFrame;
+        private Vector3 cueLocalForwardDirection;
+        private Vector3 cueArmedShotDirection;
 
         private float cueFDir;
-        protected Vector3 raySphereOutput;
-        protected int[] rackOrder8Ball = {9, 2, 10, 11, 1, 3, 4, 12, 5, 13, 14, 6, 15, 7, 8};
-        protected int[] rackOrder9Ball = {2, 3, 4, 5, 9, 6, 7, 8, 1};
-        protected int[] breakRows9ball = {0, 1, 2, 1, 0};
+        private Vector3 raySphereOutput;
+        private int[] rackOrder8Ball = {9, 2, 10, 11, 1, 3, 4, 12, 5, 13, 14, 6, 15, 7, 8};
+        private int[] rackOrder9Ball = {2, 3, 4, 5, 9, 6, 7, 8, 1};
+        private int[] breakRows9ball = {0, 1, 2, 1, 0};
 
         private float ballShadowOffset;
         private MeshRenderer[] shadowRenders;
@@ -423,6 +415,8 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
 
         private UInt32 oldDesktopCue;
         private UInt32 newDesktopCue;
+        
+        private bool fourBallCueLeftTable;
 
         /// <summary>
         /// Have we run a network sync once? Used for situations where we need to specifically catch up a late-joiner.
@@ -432,47 +426,42 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         /// <summary>
         /// For clamping to table or set lower for kitchen
         /// </summary>
-        protected float repoMaxX;
-
-        private float desktopClampX;
-        private float desktopClampY;
+        private float repoMaxX;
 
         public CueBallOffTableController cueBallController;
+        
+        private Vector3 desktopCameraInitialPosition;
+        private Quaternion desktopCameraInitialRotation;
+        private float lastLookHorizontal;
+        private float lastLookVertical;
+        private bool lastInputUseDown;
+        private float inputHeldDownTime;
+        private float desktopShootForce;
+        private const float desktopShotPowerMult = 0.15f;
+        public GameObject powerBar;
+        public GameObject topBar;
+        public Vector3 initialPowerBarPos;
 
-        private int PlayerCount
-        {
-            get
-            {
-                int count = 0;
-                if (player1ID > 0) count++;
-                if (player2ID > 0) count++;
-                if (player3ID > 0) count++;
-                if (player4ID > 0) count++;
-                return count;
-            }
-        }
-
-        public virtual void Start()
+        public void Start()
         {
             repoMaxX = tableWidth;
-            desktopClampX = tableWidth;
-            desktopClampY = tableHeight;
 
             localPlayer = Networking.LocalPlayer;
             networkingLocalPlayerID = localPlayer.playerId;
             isPlayerInVR = localPlayer.IsUserInVR();
             tableMaterials = tableRenderer.materials;
 
-            ballsArePocketed = new bool[ballTransforms.Length];
-            oldBallsArePocketed = new bool[ballTransforms.Length];
-            ballRigidbodies = new Rigidbody[ballTransforms.Length];
-            for (int i = 0; i < ballRigidbodies.Length; i++)
+            ballsArePocketed = new bool[NUMBER_OF_SIMULATED_BALLS];
+            oldBallsArePocketed = new bool[NUMBER_OF_SIMULATED_BALLS];
+            
+            ballRigidbodies = new Rigidbody[NUMBER_OF_SIMULATED_BALLS];
+            for (var i = 0; i < NUMBER_OF_SIMULATED_BALLS; i++)
             {
                 ballRigidbodies[i] = ballTransforms[i].GetComponent<Rigidbody>();
             }
 
-            ballShadowPosConstraintTransforms = new Transform[ballShadowPosConstraints.Length];
-            for (int i = 0; i < ballShadowPosConstraints.Length; i++)
+            ballShadowPosConstraintTransforms = new Transform[NUMBER_OF_SIMULATED_BALLS];
+            for (var i = 0; i < NUMBER_OF_SIMULATED_BALLS; i++)
             {
                 ballShadowPosConstraintTransforms[i] = ballShadowPosConstraints[i].transform;
             }
@@ -484,7 +473,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
                 ballPool = audioSourcePoolContainer.GetComponentsInChildren<AudioSource>();
                 ballPoolTransforms = new Transform[ballPool.Length];
 
-                for (int i = 0; i < ballPool.Length; i++)
+                for (var i = 0; i < ballPool.Length; i++)
                 {
                     ballPoolTransforms[i] = ballPool[i].transform;
                 }
@@ -516,29 +505,19 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             }
 
             if (guideline)
-            {
                 guideline.gameObject.SetActive(false);
-            }
 
             if (devhit)
-            {
                 devhit.SetActive(false);
-            }
 
             if (marker)
-            {
                 marker.SetActive(false);
-            }
 
             if (marker9ball)
-            {
                 marker9ball.SetActive(false);
-            }
 
             if (shadows)
-            {
                 shadows.SetActive(fakeBallShadows);
-            }
 
             ballShadowOffset = ballTransforms[0].position.y - ballShadowPosConstraintTransforms[0].position.y;
 
@@ -551,44 +530,32 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             desktopCameraInitialPosition = desktopCamera.transform.localPosition;
             desktopCameraInitialRotation = desktopCamera.transform.localRotation;
             initialPowerBarPos = powerBar.transform.localPosition;
+
+            CalculateTableCollisionConstants();
         }
 
-        public virtual void Update()
+        public void Update()
         {
             if (isInDesktopTopDownView)
             {
                 HandleUpdatingDesktopViewUI();
 
                 if (Input.GetKeyDown(KeyCode.E) || Input.GetKeyDown(KeyCode.Escape))
-                {
                     OnDesktopTopDownViewExit();
-                }
             }
             else if (canEnterDesktopTopDownView)
             {
                 if (Input.GetKeyDown(KeyCode.E) || lastInputUseDown)
-                {
                     OnDesktopTopDownViewStart();
-                }
             }
 
             if (isGameInMenus)
-            {
                 return;
-            }
 
             // Everything below this line only runs when the game is active.
 
-            for (int i = 0; i < ballsArePocketed.Length; i++)
-            {
-                if (!ballsArePocketed[i])
-                {
-                    ballTransforms[i].localPosition = currentBallPositions[i];
-                }
-            }
-
             localSpacePositionOfCueTip = tableSurface.transform.InverseTransformPoint(cueTip.transform.position);
-            Vector3 copyOfLocalSpacePositionOfCueTip = localSpacePositionOfCueTip;
+            var copyOfLocalSpacePositionOfCueTip = localSpacePositionOfCueTip;
 
             // if shot is prepared for next hit
             if (isPlayerAllowedToPlay)
@@ -597,48 +564,27 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
                 {
                     HandleRepositioningCueBall();
 
-                    if (IsCueContacting())
-                    {
-                        markerMaterial.SetColor(uniformMarkerColour, markerNotOK);
-                    }
-                    else
-                    {
-                        markerMaterial.SetColor(uniformMarkerColour, markerOK);
-                    }
+                    markerMaterial.SetColor(uniformMarkerColour, IsCueContacting() ? markerNotOK : markerOK);
                 }
 
-                Vector3 cueballPosition = currentBallPositions[0];
+                var cueballPosition = currentBallPositions[0];
 
                 if (isArmed)
-                {
-                    copyOfLocalSpacePositionOfCueTip =
-                        AimAndHitCueBall(copyOfLocalSpacePositionOfCueTip, cueballPosition);
-                }
+                    copyOfLocalSpacePositionOfCueTip = AimAndHitCueBall(copyOfLocalSpacePositionOfCueTip, cueballPosition);
                 else
-                {
                     HandleGuidelinesAndAimMarkers(copyOfLocalSpacePositionOfCueTip, cueballPosition);
-                }
             }
 
             localSpacePositionOfCueTipLastFrame = copyOfLocalSpacePositionOfCueTip;
 
             // Table outline colour
             if (isGameInMenus)
-            {
-                // Flashing if we won
                 tableCurrentColour = tableSrcColour * ((Mathf.Sin(Time.timeSinceLevelLoad * 3.0f) * 0.5f) + 1.0f);
-            }
             else
-            {
                 tableCurrentColour = Color.Lerp(tableCurrentColour, tableSrcColour, Time.deltaTime * 3.0f);
-            }
-
-            //float timePercentage;
-
+            
             if (isTimerRunning)
-            {
                 HandleTimerCountdown();
-            }
             else
             {
                 timerText.text = "";
@@ -652,12 +598,32 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
 
             // Run the intro animation. Do not run the animation if this is our first sync!
             if (hasRunSyncOnce && introAnimTimer > 0.0f)
-            {
                 HandleIntroAnimation();
+
+            // Run sim only if things are moving
+            if (turnIsRunning)
+            {
+                accumulatedTime += Time.deltaTime;
+
+                if (accumulatedTime > MAX_SIMULATION_TIME_PER_FRAME)
+                    accumulatedTime = MAX_SIMULATION_TIME_PER_FRAME;
+
+                while (accumulatedTime >= TIME_PER_STEP)
+                {
+                    AdvancePhysicsStep();
+                    accumulatedTime -= TIME_PER_STEP;
+                }
+            }
+            
+            if (IsCueInPlay)
+                ballTransforms[0].localPosition = currentBallPositions[0];
+            
+            for (var i = 1; i < ballsArePocketed.Length; i++)
+            {
+                if (!ballsArePocketed[i])
+                    ballTransforms[i].localPosition = currentBallPositions[i];
             }
         }
-
-        protected abstract bool IsCueContacting();
 
         private void HandleRepositioningCueBall()
         {
@@ -675,7 +641,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
 
         public void _ReEnableShadowConstraints()
         {
-            foreach (PositionConstraint con in ballShadowPosConstraints)
+            foreach (var con in ballShadowPosConstraints)
             {
                 con.constraintActive = true;
             }
@@ -708,108 +674,6 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             ReadNetworkData();
         }
 
-        public void _UnlockTable()
-        {
-            if (logger)
-            {
-                logger._Log(name, "UnlockTable");
-            }
-
-            Networking.SetOwner(localPlayer, gameObject);
-
-            isTableLocked = false;
-            RefreshNetworkData(false);
-        }
-
-        public void _LockTable()
-        {
-            if (logger)
-            {
-                logger._Log(name, "LockTable");
-            }
-
-            Networking.SetOwner(localPlayer, gameObject);
-
-            isTableLocked = true;
-            RefreshNetworkData(false);
-        }
-
-        public void _JoinGame(int playerNumber)
-        {
-            if (logger)
-            {
-                logger._Log(name, $"JoinGame: {playerNumber}");
-            }
-
-            Networking.SetOwner(localPlayer, gameObject);
-            localPlayerID = playerNumber;
-
-            switch (playerNumber)
-            {
-                case 0:
-                    player1ID = networkingLocalPlayerID;
-                    EnableCustomBallColorSlider(true);
-                    break;
-                case 1:
-                    player2ID = networkingLocalPlayerID;
-                    EnableCustomBallColorSlider(true);
-                    break;
-                case 2:
-                    player3ID = networkingLocalPlayerID;
-                    EnableCustomBallColorSlider(true);
-                    break;
-                case 3:
-                    player4ID = networkingLocalPlayerID;
-                    EnableCustomBallColorSlider(true);
-                    break;
-                default:
-                    return;
-            }
-
-            RefreshNetworkData(false);
-
-            if (slogger)
-            {
-                slogger.OscReportJoinedTeam(playerNumber);
-            }
-        }
-
-        public void _LeaveGame()
-        {
-            if (logger)
-            {
-                logger._Log(name, "LeaveGame");
-            }
-
-            Networking.SetOwner(localPlayer, gameObject);
-
-            switch (localPlayerID)
-            {
-                case 0:
-                    player1ID = 0;
-                    break;
-                case 1:
-                    player2ID = 0;
-                    break;
-                case 2:
-                    player3ID = 0;
-                    break;
-                case 3:
-                    player4ID = 0;
-                    break;
-                default:
-                    return;
-            }
-
-            localPlayerID = -1;
-
-            RefreshNetworkData(false);
-
-            //akalink added, makes the color panel not able to be interacted with
-            EnableCustomBallColorSlider(false);
-            //end
-        }
-
         // Update loop-scoped handler for introduction animation operations. Non-pure.
         private void HandleIntroAnimation()
         {
@@ -820,9 +684,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             float aitime;
 
             if (introAnimTimer < 0.0f)
-            {
                 introAnimTimer = 0.0f;
-            }
 
             // Cueball drops late
             Transform ball = ballTransforms[0];
@@ -895,100 +757,69 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
                     OnTurnOverFoul();
                 }
                 else
-                {
-                    // All local players freeze until next target
-                    // can pick up and propogate timer end
                     isPlayerAllowedToPlay = false;
-                }
-
-                //timePercentage = 0.0f;
             }
             else
             {
                 if (timerText)
-                {
                     timerText.text = timerOutputFormat.Replace("{}", Mathf.Round(remainingTime).ToString());
-                }
 
                 if (timerCountdown)
-                {
                     timerCountdown.fillAmount = remainingTime / timerSecondsPerShot;
-                }
             }
         }
-
-        protected abstract bool IsIntersectingWithSphere(Vector3 start, Vector3 dir, Vector3 sphere);
 
         // Update loop-scoped handler for guidelines and aim markers. Non-pure.
         private void HandleGuidelinesAndAimMarkers(Vector3 copyOfLocalSpacePositionOfCueTip, Vector3 cueballPosition)
         {
             cueLocalForwardDirection = tableSurface.transform.InverseTransformVector(cueTip.transform.forward);
+            
+            if (devhit)
+                devhit.SetActive(false);
+
+            if (guideline)
+                guideline.gameObject.SetActive(false);
 
             // Get where the cue will strike the ball
-            if (IsIntersectingWithSphere(copyOfLocalSpacePositionOfCueTip, cueLocalForwardDirection,
-                    cueballPosition))
+            if (!IsIntersectingWithSphere(copyOfLocalSpacePositionOfCueTip, cueLocalForwardDirection, cueballPosition)) 
+
+                return;
+            if (guideLineEnabled && guideline)
+                guideline.gameObject.SetActive(true);
+
+            if (devhit)
             {
-                if (guideLineEnabled && guideline)
-                {
-                    guideline.gameObject.SetActive(true);
-                }
-
-                if (devhit)
-                {
-                    devhit.SetActive(true);
-                    devhit.transform.localPosition = raySphereOutput;
-                }
-
-                cueArmedShotDirection = cueLocalForwardDirection;
-                // cueArmedShotDirection.y = 0.0f;
-
-                if (!isInDesktopTopDownView)
-                {
-                    // Compute deflection in VR mode
-                    Vector3 scuffdir = cueballPosition - raySphereOutput;
-                    //scuffdir.y = 0.0f;
-                    cueArmedShotDirection += scuffdir.normalized * 0.17f;
-                }
-
-                cueFDir = Mathf.Atan2(cueArmedShotDirection.z, cueArmedShotDirection.x);
-
-                // Update the prediction line direction
-                Transform transform1 = guideline.transform;
-                transform1.localPosition = currentBallPositions[0];
-                transform1.localEulerAngles = new Vector3(0.0f, -cueFDir * Mathf.Rad2Deg, 0.0f);
+                devhit.SetActive(true);
+                devhit.transform.localPosition = raySphereOutput;
             }
-            else
+
+            cueArmedShotDirection = cueLocalForwardDirection;
+
+            if (!isInDesktopTopDownView)
             {
-                if (devhit)
-                {
-                    devhit.SetActive(false);
-                }
-
-                if (guideline)
-                {
-                    guideline.gameObject.SetActive(false);
-                }
+                // Compute deflection in VR mode
+                Vector3 scuffdir = cueballPosition - raySphereOutput;
+                cueArmedShotDirection += scuffdir.normalized * 0.17f;
             }
+
+            cueFDir = Mathf.Atan2(cueArmedShotDirection.z, cueArmedShotDirection.x);
+
+            // Update the prediction line direction
+            Transform transform1 = guideline.transform;
+            transform1.localPosition = currentBallPositions[0];
+            transform1.localEulerAngles = new Vector3(0.0f, -cueFDir * Mathf.Rad2Deg, 0.0f);
         }
-
-        protected abstract Vector3 AimAndHitCueBall(Vector3 copyOfLocalSpacePositionOfCueTip, Vector3 cueballPosition);
 
         private void EnableCustomBallColorSlider(bool enabledState)
         {
             if (!ballCustomColours)
-            {
                 return;
-            }
 
             if (blueTeamSliders)
-            {
                 blueTeamSliders._EnableDisable(enabledState);
-            }
 
             if (orangeTeamSliders)
-            {
                 orangeTeamSliders._EnableDisable(enabledState);
-            }
         }
 
         private void RemovePlayerFromGame(int playerID)
@@ -1016,302 +847,10 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             RefreshNetworkData(false);
         }
 
-        public void _IncreaseTimer()
-        {
-            if (logger)
-            {
-                logger._Log(name, "IncreaseTimer");
-            }
-
-            Networking.SetOwner(localPlayer, gameObject);
-            timerSecondsPerShot += 5;
-            RefreshNetworkData(false);
-
-            if (timerSecondsPerShot >= 60)
-            {
-                timerSecondsPerShot = 60;
-            }
-        }
-
-        public void _DecreaseTimer()
-        {
-            if (logger)
-            {
-                logger._Log(name, "DecreaseTimer");
-            }
-
-            Networking.SetOwner(localPlayer, gameObject);
-            timerSecondsPerShot -= 5;
-
-            if (timerSecondsPerShot <= 0)
-            {
-                timerSecondsPerShot = 0;
-            }
-
-            RefreshNetworkData(false);
-        }
-
-        public void _SelectTeams()
-        {
-            if (logger)
-            {
-                logger._Log(name, "SelectTeams");
-            }
-
-            Networking.SetOwner(localPlayer, gameObject);
-
-            isTeams = true;
-            RefreshNetworkData(false);
-        }
-
-        public void _DeselectTeams()
-        {
-            if (logger)
-            {
-                logger._Log(name, "DeselectTeams");
-            }
-
-            Networking.SetOwner(localPlayer, gameObject);
-
-            isTeams = false;
-            RefreshNetworkData(false);
-        }
-
-        public void _EnableGuideline()
-        {
-            if (logger)
-            {
-                logger._Log(name, "EnableGuideline");
-            }
-
-            Networking.SetOwner(localPlayer, gameObject);
-            guideLineEnabled = true;
-            RefreshNetworkData(false);
-        }
-
-        public void _DisableGuideline()
-        {
-            if (logger)
-            {
-                logger._Log(name, "DisableGuideline");
-            }
-
-            Networking.SetOwner(localPlayer, gameObject);
-            guideLineEnabled = false;
-            RefreshNetworkData(false);
-        }
-
-
-        /// <summary>
-        /// Initialize new match as the host.
-        /// </summary>
-        public void _StartNewGame()
-        {
-            if (logger)
-            {
-                logger._Log(name, "StartNewGame");
-            }
-
-            mainSrc.enabled = true;
-
-            if (!isGameInMenus)
-            {
-                return;
-            }
-
-            gameWasReset = false;
-            gameID++;
-            turnID = 0;
-            isPlayerAllowedToPlay = true;
-
-            isTeam2Turn = false;
-            oldIsTeam2Turn = false;
-
-            // Following is overrides of NewGameLocal, for game STARTER only
-            turnIsRunning = false;
-            isOpen = true;
-            isGameInMenus = false;
-            poolCues[0].tableIsActive = true;
-            poolCues[1].tableIsActive = true;
-
-            isTeam2Blue = false;
-            isTeam2Winner = false;
-            isFoul = false;
-            isGameOver = false;
-
-            ApplyTableColour(false);
-
-            Networking.SetOwner(localPlayer, gameObject);
-
-            RefreshNetworkData(false);
-        }
-
-
-        public void _Select8Ball()
-        {
-            if (logger)
-            {
-                logger._Log(name, "Select8Ball");
-            }
-
-            Networking.SetOwner(localPlayer, gameObject);
-
-            gameMode = 0u;
-            RefreshNetworkData(false);
-        }
-
-        public void _Select9Ball()
-        {
-            if (logger)
-            {
-                logger._Log(name, "Select9Ball");
-            }
-
-            Networking.SetOwner(localPlayer, gameObject);
-
-            gameMode = 1u;
-            RefreshNetworkData(false);
-        }
-
-        public void _Select4BallJapanese()
-        {
-            if (logger)
-            {
-                logger._Log(name, "Select4BallJapanese");
-            }
-
-            Networking.SetOwner(localPlayer, gameObject);
-
-            isKorean = false;
-            isThreeCushionCarom = false;
-
-            gameMode = 2u;
-            RefreshNetworkData(false);
-        }
-
-        public void _SelectThreeCushionCarom()
-        {
-            if (logger)
-            {
-                logger._Log(name, "SelectThreeCushionCarom");
-            }
-
-            Networking.SetOwner(localPlayer, gameObject);
-
-            isKorean = false;
-            isThreeCushionCarom = true;
-
-            gameMode = 2u;
-            RefreshNetworkData(false);
-        }
-
-
-        public void _Select4BallKorean()
-        {
-            if (logger)
-            {
-                logger._Log(name, "Select4BallKorean");
-            }
-
-            Networking.SetOwner(localPlayer, gameObject);
-
-            isKorean = true;
-            isThreeCushionCarom = false;
-
-            gameMode = 2u;
-            RefreshNetworkData(false);
-        }
-
-        private void Initialize8Ball()
-        {
-            ballsArePocketed = new[]
-            {
-                false, false, false, false, false, false, false, false, false, false, false, false, false, false, false,
-                false
-            };
-
-            for (int i = 0, k = 0; i < 5; i++)
-            {
-                for (int j = 0; j <= i; j++)
-                {
-                    currentBallPositions[rackOrder8Ball[k++]] = new Vector3
-                    (
-                        SPOT_POSITION_X + (i * BALL_PL_Y) + UnityEngine.Random.Range(-RANDOMIZE_F, RANDOMIZE_F),
-                        0.0f,
-                        ((-i + (j * 2)) * BALL_PL_X) + UnityEngine.Random.Range(-RANDOMIZE_F, RANDOMIZE_F)
-                    );
-
-                    currentBallVelocities[k] = Vector3.zero;
-                    currentAngularVelocities[k] = Vector3.zero;
-                }
-            }
-
-            TeamColors();
-        }
-
-        private void Initialize9Ball()
-        {
-            //ballPocketedState = 0xFC00u;
-            ballsArePocketed = new[]
-            {
-                false, false, false, false, false, false, false, false, false, false, true, true, true, true, true, true
-            };
-
-            for (int i = 0, k = 0; i < 5; i++)
-            {
-                int rown = breakRows9ball[i];
-                for (int j = 0; j <= rown; j++)
-                {
-                    currentBallPositions[rackOrder9Ball[k++]] = new Vector3
-                    (
-                        SPOT_POSITION_X + (i * BALL_PL_Y) + UnityEngine.Random.Range(-RANDOMIZE_F, RANDOMIZE_F),
-                        0.0f,
-                        ((-rown + (j * 2)) * BALL_PL_X) + UnityEngine.Random.Range(-RANDOMIZE_F, RANDOMIZE_F)
-                    );
-
-                    currentBallVelocities[k] = Vector3.zero;
-                    currentAngularVelocities[k] = Vector3.zero;
-                }
-            }
-
-            UsColors();
-        }
-
-        private Vector3 GetFourBallCueStartingPosition(bool team)
-        {
-            return team ? new Vector3(SPOT_CAROM_X, 0.0f, 0.0f) : new Vector3(-SPOT_CAROM_X, 0.0f, 0.0f);
-        }
-
-        private void Initialize4Ball()
-        {
-            //ballPocketedState = 0xFDF2u;
-            ballsArePocketed = new[]
-                {false, true, false, false, true, true, true, true, true, false, true, true, true, true, true, true};
-
-            currentBallPositions[0] = new Vector3(-SPOT_CAROM_X, 0.0f, 0.0f);
-            currentBallPositions[9] = new Vector3(SPOT_CAROM_X, 0.0f, 0.0f);
-            currentBallPositions[2] = new Vector3(SPOT_POSITION_X, 0.0f, 0.0f);
-            currentBallPositions[3] = new Vector3(-SPOT_POSITION_X, 0.0f, 0.0f);
-
-            currentBallVelocities[0] = Vector3.zero;
-            currentBallVelocities[9] = Vector3.zero;
-            currentBallVelocities[2] = Vector3.zero;
-            currentBallVelocities[3] = Vector3.zero;
-
-            currentAngularVelocities[0] = Vector3.zero;
-            currentAngularVelocities[9] = Vector3.zero;
-            currentAngularVelocities[2] = Vector3.zero;
-            currentAngularVelocities[3] = Vector3.zero;
-
-            FourBallColors();
-        }
-
         public void _StartHit()
         {
             if (logger)
-            {
                 logger._Log(name, "StartHit");
-            }
 
             // lock aim variables
             bool isOurTurn = ((localPlayerID >= 0) && (playerIsTeam2 == isTeam2Turn)) || isGameModePractice;
@@ -1325,9 +864,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         public void _EndHit()
         {
             if (logger)
-            {
                 logger._Log(name, "EndHit");
-            }
 
             isArmed = false;
         }
@@ -1335,68 +872,26 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         public void PlaceBall()
         {
             if (logger)
-            {
                 logger._Log(name, "PlaceBall");
-            }
 
-            if (!IsCueContacting())
-            {
-                if (logger)
-                {
-                    logger._Log(name, "disabling marker because the ball hase been placed");
-                }
+            if (IsCueContacting()) 
+                return;
 
-                isRepositioningCueBall = false;
-
-                if (marker)
-                {
-                    marker.SetActive(false);
-                }
-
-                isPlayerAllowedToPlay = true;
-                isFoul = false;
-
-                Networking.SetOwner(localPlayer, gameObject);
-
-                // Save out position to remote clients
-                RefreshNetworkData(isTeam2Turn);
-            }
-        }
-
-        public void _ForceReset()
-        {
             if (logger)
-            {
-                logger._Log(name, "ForceReset");
-            }
+                logger._Log(name, "disabling marker because the ball hase been placed");
 
-            if (Networking.IsInstanceOwner)
-            {
-                Networking.SetOwner(localPlayer, gameObject);
-                _Reset(ResetReason.InstanceOwnerReset);
-            }
-            else if (
-                networkingLocalPlayerID == player1ID || networkingLocalPlayerID == player2ID ||
-                networkingLocalPlayerID == player3ID || networkingLocalPlayerID == player4ID
-            )
-            {
-                Networking.SetOwner(localPlayer, gameObject);
-                _Reset(ResetReason.PlayerReset);
-            }
-            else if (
-                (player1ID > 0 && !VRCPlayerApi.GetPlayerById(player1ID).IsValid()) ||
-                (player2ID > 0 && !VRCPlayerApi.GetPlayerById(player2ID).IsValid()) ||
-                (player3ID > 0 && !VRCPlayerApi.GetPlayerById(player3ID).IsValid()) ||
-                (player4ID > 0 && !VRCPlayerApi.GetPlayerById(player4ID).IsValid())
-            )
-            {
-                Networking.SetOwner(localPlayer, gameObject);
-                _Reset(ResetReason.InvalidState);
-            }
-            else if (logger)
-            {
-                logger._Error(name, "Cannot reset table: you do not have permission");
-            }
+            isRepositioningCueBall = false;
+
+            if (marker)
+                marker.SetActive(false);
+
+            isPlayerAllowedToPlay = true;
+            isFoul = false;
+
+            Networking.SetOwner(localPlayer, gameObject);
+
+            // Save out position to remote clients
+            RefreshNetworkData(isTeam2Turn);
         }
 
         private void _Reset(ResetReason reason)
@@ -1414,17 +909,13 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             RefreshNetworkData(false);
 
             if (logger)
-            {
                 logger._Log(name, ToReasonString(reason));
-            }
         }
 
         private void OnDesktopTopDownViewStart()
         {
             if (logger)
-            {
                 logger._Log(name, "OnDesktopTopDownViewStart");
-            }
 
             isInDesktopTopDownView = true;
             isEnteringDesktopModeThisFrame = true;
@@ -1448,17 +939,13 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         public void _OnPutDownCueLocally()
         {
             if (logger)
-            {
                 logger._Log(name, "OnPutDownCueLocally");
-            }
 
             OnDesktopTopDownViewExit();
         }
 
-        protected bool fourBallCueLeftTable;
-
         // HandleEndOfTurn assess the table state at the end of the turn, based on what game is playing.
-        protected void HandleEndOfTurn()
+        private void HandleEndOfTurn()
         {
             isSimulatedByUs = false;
 
@@ -1474,126 +961,83 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             var numberOfSunkBlues = 0;
             var numberOfSunkOranges = 0;
 
-            if (is8Ball) // Standard 8 ball
-            {
-                var isBlue = !isOpen && (isTeam2Turn && isTeam2Blue || !isTeam2Turn && !isTeam2Blue);
+            switch (gameMode) {
+                case GameMode.EightBall:
+                    HandleEightBallEndOfTurn(
+                        is8Sink,
+                        out numberOfSunkBlues,
+                        out numberOfSunkOranges,
+                        ref isCorrectBallSunk,
+                        ref isOpponentColourSunk,
+                        ref foulCondition,
+                        out deferLossCondition,
+                        out winCondition
+                    );
 
-                numberOfSunkBlues = GetNumberOfSunkBlues();
-                numberOfSunkOranges = GetNumberOfSunkOranges();
-                var isWrongHit = IsFirstEightBallHitFoul(firstHitBallThisTurn, numberOfSunkBlues, numberOfSunkOranges);
+                    break;
+                case GameMode.NineBall:
+                    HandleNineBallEndOfTurn(ref foulCondition, ref isCorrectBallSunk, out winCondition);
 
-                // What balls got sunk this turn?
-                for (var i = 2; i < NUMBER_OF_SIMULATED_BALLS; i++)
-                {
-                    if (ballsArePocketed[i] == oldBallsArePocketed[i])
-                        continue;
+                    break;
+                case GameMode.KoreanCarom:
+                case GameMode.JapaneseCarom:
+                    HandleCaromEndOfTurn();
 
-                    if (isOpen)
-                        isCorrectBallSunk = true;
-                    else if (isBlue)
-                    {
-                        if (i > 1 && i < 9)
-                            isCorrectBallSunk = true;
-                        else
-                            isOpponentColourSunk = true;
-                    }
-                    else
-                    {
-                        if (i >= 9)
-                            isCorrectBallSunk = true;
-                        else
-                            isOpponentColourSunk = true;
-                    }
-                }
+                    isCorrectBallSunk = isMadePoint;
+                    isOpponentColourSunk = isMadeFoul;
+                    winCondition = scores[Convert.ToInt32(isTeam2Turn)] >= scoreNeededToWinCarom;
+                    
+                    break;
+                case GameMode.ThreeCushionCarom:
+                    HandleThreeCushionCaromEndOfTurn();
+                    
+                    // There's no action to take when fouling in 3CC (no marker spawn etc) so we just care if the player
+                    // scored.
+                    isCorrectBallSunk = isMadePoint;
+                    winCondition = scores[Convert.ToInt32(isTeam2Turn)] >= scoreNeededToWinCarom;
 
-                winCondition = isBlue ? numberOfSunkBlues == 7 && is8Sink : numberOfSunkOranges == 7 && is8Sink;
-
-                if (isWrongHit)
-                {
-                    if (logger)
-                        logger._Log(name, "foul, wrong ball hit");
-
-                    foulCondition = true;
-                }
-                else if (firstHitBallThisTurn == 0)
-                {
-                    if (logger)
-                        logger._Log(name, "foul, no ball hit");
-
-                    foulCondition = true;
-                }
-
-                deferLossCondition = is8Sink;
+                    break;
+                default:
+                    return;
             }
-            else if (isNineBall) // 9 ball
-            {
-                var isWrongHit = GetLowestNumberedBall(oldBallsArePocketed) != firstHitBallThisTurn;
-                foulCondition = foulCondition || isWrongHit || firstHitBallThisTurn == 0;
 
-                // Win condition: Pocket 9 ball ( at anytime )
-                winCondition = ballsArePocketed[9];
-
-                // this video is hard to follow so im just gonna guess this is right
-                for (int i = 1; i < 10; i++)
-                {
-                    if (ballsArePocketed[i] != oldBallsArePocketed[i])
-                    {
-                        isCorrectBallSunk = true;
-                    }
-                }
-
-                // TODO: Implement rail contact requirement
-            }
-            else if (isFourBall) // 4 ball
-            {
-                if (fourBallCueLeftTable)
-                {
-                    currentBallPositions[0] = GetFourBallCueStartingPosition(isTeam2Turn);
-                    currentBallVelocities[0] = Vector3.zero;
-                    currentAngularVelocities[0] = Vector3.zero;
-
-                    // Best effort attempt to place the ball somewhere
-                    if (IsCueContacting())
-                    {
-                        currentBallPositions[0] = GetFourBallCueStartingPosition(!isTeam2Turn);
-
-                        if (IsCueContacting())
-                        {
-                            currentBallPositions[0] = Vector3.zero;
-
-                            // Let the player fix it
-                            isFoul = true;
-                        }
-                    }
-
-                    fourBallCueLeftTable = false;
-                }
-
-                isCorrectBallSunk = isMadePoint;
-                isOpponentColourSunk = isMadeFoul;
-                winCondition = scores[Convert.ToInt32(isTeam2Turn)] >= scoreNeededToWinCarom;
-            }
-            else
-                return;
-
+            // Has the player won?
             if (winCondition)
             {
+                // Has the player fouled, giving the opponent the win?
                 if (foulCondition)
                     OnTurnOverGameWon(!isTeam2Turn, true);
                 else
                     OnTurnOverGameWon(isTeam2Turn, false);
+
+                return;
             }
-            else if (deferLossCondition)
-                OnTurnOverGameWon(!isTeam2Turn, true);
-            else if (foulCondition)
-                OnTurnOverFoul();
-            else if (isCorrectBallSunk && !isOpponentColourSunk)
+            
+            // Has the player fouled catastrophically and lost?
+            if (deferLossCondition)
             {
-                if (is8Ball && isOpen)
+                OnTurnOverGameWon(!isTeam2Turn, true);
+
+                return;
+            }
+
+            // Has the player fouled?
+            if (foulCondition)
+            {
+                OnTurnOverFoul();
+
+                return;
+            }
+            
+            // Has the player done enough to continue their turn?
+            if (isCorrectBallSunk && !isOpponentColourSunk)
+            {
+                // TODO: not happy with this, revise this based on blackball rules.
+                if (gameMode == GameMode.EightBall && isOpen)
                 {
                     if (numberOfSunkBlues != numberOfSunkOranges)
                     {
-                        isTeam2Blue = (numberOfSunkBlues > numberOfSunkOranges) ? isTeam2Turn : !isTeam2Turn;
+                        isTeam2Blue = numberOfSunkBlues > numberOfSunkOranges ? isTeam2Turn : !isTeam2Turn;
 
                         isOpen = false;
                         ApplyTableColour(isTeam2Turn);
@@ -1601,56 +1045,30 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
                 }
 
                 isPlayerAllowedToPlay = true;
+                _isGameBreak = false;
 
                 RefreshNetworkData(isTeam2Turn);
+
+                return;
             }
-            else
+            
+            // End the player's turn and pass the turn on.
+            
+            isPlayerAllowedToPlay = true;
+            _isGameBreak = false;
+
+            if (IsCarom)
             {
-                isPlayerAllowedToPlay = true;
-
-                if (isFourBall)
-                {
-                    Vector3 temp = currentBallPositions[0];
-                    currentBallPositions[0] = currentBallPositions[9];
-                    currentBallPositions[9] = temp;
-                }
-
-                turnID++;
-
-                RefreshNetworkData(!isTeam2Turn);
-            }
-        }
-
-        private int GetNumberOfSunkOranges()
-        {
-            var num = 0;
-
-            for (var i = 9; i < 16; i++)
-            {
-                if (ballsArePocketed[i])
-                {
-                    num++;
-                }
+                Vector3 temp = currentBallPositions[0];
+                currentBallPositions[0] = currentBallPositions[9];
+                currentBallPositions[9] = temp;
             }
 
-            return num;
+            turnID++;
+
+            RefreshNetworkData(!isTeam2Turn);
         }
-
-        private int GetNumberOfSunkBlues()
-        {
-            var num = 0;
-
-            for (var i = 2; i < 9; i++)
-            {
-                if (ballsArePocketed[i])
-                {
-                    num++;
-                }
-            }
-
-            return num;
-        }
-
+        
         private void RefreshNetworkData(bool newIsTeam2Playing)
         {
             if (logger)
@@ -1665,20 +1083,16 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             ReadNetworkData();
         }
 
-        protected void ReadNetworkData()
+        private void ReadNetworkData()
         {
             if (logger)
-            {
                 logger._Log(name, "ReadNetworkData");
-            }
 
             UpdateScores();
 
             // Assume the marker is off - one of the On... functions might turn it back on again
             if (marker)
-            {
                 marker.SetActive(false);
-            }
 
             if (gameID > oldGameID && !isGameInMenus)
                 OnNewGameStarted();
@@ -1702,10 +1116,8 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             poolMenu._UpdateMainMenuView(
                 isTeams,
                 isTeam2Turn,
-                (int) gameMode,
-                isKorean,
-                isThreeCushionCarom,
-                (int) timerSecondsPerShot,
+                gameMode,
+                timerSecondsPerShot,
                 player1ID,
                 player2ID,
                 player3ID,
@@ -1720,24 +1132,16 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
                 if (!isTableLocked)
                 {
                     if (player1ID != 0)
-                    {
                         numberOfPlayers++;
-                    }
 
                     if (player2ID != 0)
-                    {
                         numberOfPlayers++;
-                    }
 
                     if (player3ID != 0)
-                    {
                         numberOfPlayers++;
-                    }
 
                     if (player4ID != 0)
-                    {
                         numberOfPlayers++;
-                    }
                 }
 
                 isGameModePractice = localPlayerID == 0 && numberOfPlayers == 1;
@@ -1747,24 +1151,27 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
                 return;
             }
 
-            if (isFourBall)
+            switch (gameMode)
             {
-                ballsArePocketed = new[]
-                {
-                    false, true, false, false, true, true, true, true, true, false, true, true, true, true, true, true
-                };
+                case GameMode.KoreanCarom:
+                case GameMode.JapaneseCarom:
+                    InitializePocketedStateKoreanJapaneseCarom();
+                    break;
+                case GameMode.ThreeCushionCarom:
+                    InitializePocketedStateThreeCushionCarom();
+                    break;
+                case GameMode.EightBall:
+                case GameMode.NineBall:
+                default:
+                    break;
             }
 
             // Check this every read
             // Its basically 'turn start' event
             if (isPlayerAllowedToPlay)
-            {
                 OnLocalTurnStart();
-            }
             else
-            {
                 OnRemoteTurnStart();
-            }
 
             // Start of turn so we've hit nothing
             firstHitBallThisTurn = 0;
@@ -1778,23 +1185,17 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         private void OnRemoteTurnStart()
         {
             if (marker9ball)
-            {
                 marker9ball.SetActive(false);
-            }
 
             isTimerRunning = false;
             isMadePoint = false;
             isMadeFoul = false;
 
             if (devhit)
-            {
                 devhit.SetActive(false);
-            }
 
             if (guideline)
-            {
                 guideline.gameObject.SetActive(false);
-            }
         }
 
         private void OnLocalTurnStart()
@@ -1808,11 +1209,9 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
                 desktopHitPoint = Vector3.zero;
             }
             else
-            {
                 isDesktopLocalTurn = false;
-            }
 
-            if (isNineBall)
+            if (gameMode == GameMode.NineBall)
             {
                 int target = GetLowestNumberedBall(ballsArePocketed);
 
@@ -1826,18 +1225,13 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             }
 
             if (!tableModelHasRails || !hasRunSyncOnce)
-            {
                 PlaceSunkBallsIntoRestingPlace();
-            }
 
             if (timerSecondsPerShot > 0 && !isTimerRunning)
-            {
                 ResetTimer();
-            }
 
             // sanitize old cue tip location data to prevent stale data from causing unintended effects.
-            localSpacePositionOfCueTipLastFrame =
-                tableSurface.transform.InverseTransformPoint(cueTip.transform.position);
+            localSpacePositionOfCueTipLastFrame = tableSurface.transform.InverseTransformPoint(cueTip.transform.position);
         }
 
         private void OnNewGameStarted()
@@ -1847,36 +1241,32 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             if (((localPlayerID >= 0) && (playerIsTeam2 == isTeam2Turn)) || isGameModePractice)
             {
                 if (logger)
-                {
                     logger._Log(name, "enabling marker because it is the start of the game and we are breaking");
-                }
 
                 isRepositioningCueBall = true;
                 repoMaxX = -SPOT_POSITION_X;
                 ballRigidbodies[0].isKinematic = true;
 
-                if (marker)
-                {
-                    markerTransform.localPosition = currentBallPositions[0];
-                    if (!isFourBall)
-                    {
-                        marker.SetActive(true);
-                    }
-
-                    ((VRC_Pickup) marker.gameObject.GetComponent(typeof(VRC_Pickup))).pickupable = true;
-                }
-            }
-            else
-            {
+                if (!marker) 
+                    return;
+                
                 markerTransform.localPosition = currentBallPositions[0];
 
-                if (!isFourBall)
+                if (gameMode == GameMode.EightBall || gameMode == GameMode.NineBall)
                     marker.SetActive(true);
 
-                ((VRC_Pickup) marker.gameObject.GetComponent(typeof(VRC_Pickup))).pickupable = false;
-            }
-        }
+                ((VRC_Pickup) marker.gameObject.GetComponent(typeof(VRC_Pickup))).pickupable = true;
 
+                return;
+            }
+            
+            markerTransform.localPosition = currentBallPositions[0];
+
+            if (gameMode == GameMode.EightBall || gameMode == GameMode.NineBall)
+                marker.SetActive(true);
+
+            ((VRC_Pickup) marker.gameObject.GetComponent(typeof(VRC_Pickup))).pickupable = false;
+        }
 
         public void OnDisable()
         {
@@ -1890,6 +1280,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         {
             ballColors[0] = Color.white;
             ballColors[1] = Color.black;
+
             for (int i = 2; i < 9; i++)
             {
                 ballColors[i] = tableBlue;
@@ -1922,7 +1313,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             ballColors[15] = new Color(0.59f, 0.29f, 0, 1);
         }
 
-        private void FourBallColors() // will be used later. Doesnt do much yet
+        private void SetFourBallColours()
         {
             ballColors[0] = Color.white;
             ballColors[9] = Color.yellow;
@@ -1936,84 +1327,32 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         private void ApplyTableColour(bool isTeam2Color)
         {
             if (logger)
-            {
                 logger._Log(name, "ApplyTableColour");
-            }
 
-            if (isFourBall)
-            {
-                if (!this.isTeam2Turn)
-                {
-                    tableSrcColour = pointerColour0;
-                    cueRenderObjs[0].materials[0].SetColor(uniformCueColour, pointerColour0);
-                    cueRenderObjs[1].materials[0].SetColor(uniformCueColour, pointerColour1 * 0.5f);
-                }
-                else
-                {
-                    tableSrcColour = pointerColour1;
-                    cueRenderObjs[0].materials[0].SetColor(uniformCueColour, pointerColour0 * 0.5f);
-                    cueRenderObjs[1].materials[0].SetColor(uniformCueColour, pointerColour1);
-                }
-            }
-            else if (isNineBall)
-            {
-                int target = GetLowestNumberedBall(ballsArePocketed);
-                Color color = ballColors[target];
-                cueRenderObjs[Convert.ToInt32(isTeam2Color)].materials[0].SetColor(uniformCueColour, color);
-                cueRenderObjs[Convert.ToInt32(!isTeam2Color)].materials[0].SetColor(uniformCueColour, tableBlack);
+            switch(gameMode) {
+                case GameMode.EightBall:
+                    ApplyEightBallTableColour(isTeam2Color);
 
-                tableSrcColour = color;
-            }
-            else if (!isOpen)
-            {
-                if (isTeam2Color)
-                {
-                    if (isTeam2Blue)
-                    {
-                        tableSrcColour = tableBlue;
-                        cueRenderObjs[0].material.SetColor(uniformCueColour, tableOrange * 0.33f);
-                        cueRenderObjs[1].material.SetColor(uniformCueColour, tableBlue);
-                    }
-                    else
-                    {
-                        tableSrcColour = tableOrange;
-                        cueRenderObjs[0].material.SetColor(uniformCueColour, tableBlue * 0.33f);
-                        cueRenderObjs[1].material.SetColor(uniformCueColour, tableOrange);
-                    }
-                }
-                else
-                {
-                    if (isTeam2Blue)
-                    {
-                        tableSrcColour = tableOrange;
-                        cueRenderObjs[0].material.SetColor(uniformCueColour, tableOrange);
-                        cueRenderObjs[1].material.SetColor(uniformCueColour, tableBlue * 0.33f);
-                    }
-                    else
-                    {
-                        tableSrcColour = tableBlue;
-                        cueRenderObjs[0].material.SetColor(uniformCueColour, tableBlue);
-                        cueRenderObjs[1].material.SetColor(uniformCueColour, tableOrange * 0.33f);
-                    }
-                }
-            }
-            else
-            {
-                tableSrcColour = pointerColour2;
+                    break;
+                case GameMode.NineBall:
+                    ApplyNineBallTableColour(isTeam2Color);
 
-                cueRenderObjs[Convert.ToInt32(this.isTeam2Turn)].materials[0].SetColor(uniformCueColour, tableWhite);
-                cueRenderObjs[Convert.ToInt32(!this.isTeam2Turn)].materials[0].SetColor(uniformCueColour, tableBlack);
+                    break;
+                case GameMode.KoreanCarom:
+                case GameMode.JapaneseCarom:
+                case GameMode.ThreeCushionCarom:
+                    ApplyCaromTableColour();
+
+                    break;
             }
 
             cueGrips[Convert.ToInt32(this.isTeam2Turn)].SetColor(uniformMarkerColour, gripColourActive);
             cueGrips[Convert.ToInt32(!this.isTeam2Turn)].SetColor(uniformMarkerColour, gripColourInactive);
 
             if (!ballCustomColours)
-            {
                 return;
-            }
 
-            if (isFourBall || isNineBall)
+            if (gameMode == GameMode.NineBall)
             {
                 foreach (MeshRenderer meshRenderer in ballRenderers)
                 {
@@ -2022,7 +1361,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             }
             else
             {
-                for (var i = 2; i < ballTransforms.Length; i++)
+                for (var i = 2; i < NUMBER_OF_SIMULATED_BALLS; i++)
                 {
                     if (i < 9) //9 is where it switches to stripes
                     {
@@ -2038,30 +1377,6 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
                     }
                 }
             }
-        }
-
-        private void SpawnPlusOne(Transform ball)
-        {
-            if (logger)
-            {
-                logger._Log(name, "SpawnPlusOne");
-            }
-
-            ParticleSystem.EmitParams emitParams = new ParticleSystem.EmitParams();
-            emitParams.position = ball.position;
-            plusOneParticleSystem.Emit(emitParams, 1);
-        }
-
-        private void SpawnMinusOne(Transform ball)
-        {
-            if (logger)
-            {
-                logger._Log(name, "SpawnMinusOne");
-            }
-
-            ParticleSystem.EmitParams emitParams = new ParticleSystem.EmitParams();
-            emitParams.position = ball.position;
-            minusOneParticleSystem.Emit(emitParams, 1);
         }
 
         private void ResetTimer()
@@ -2080,21 +1395,6 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             {
                 isTimerRunning = false;
             }
-        }
-
-        private void OnLocalCaromPoint(Transform ball)
-        {
-            isMadePoint = true;
-            mainSrc.PlayOneShot(pointMadeSfx, 1.0f);
-
-            scores[Convert.ToUInt32(isTeam2Turn)]++;
-
-            if (scores[Convert.ToUInt32(isTeam2Turn)] > 10)
-            {
-                scores[Convert.ToUInt32(isTeam2Turn)] = 10;
-            }
-
-            SpawnPlusOne(ball);
         }
 
         /// <summary>
@@ -2119,9 +1419,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
                 }
 
                 if (slogger)
-                {
                     slogger.OscReportGameReset(latestResetReason);
-                }
             }
             else
             {
@@ -2130,26 +1428,18 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             }
 
             if (marker9ball)
-            {
                 marker9ball.SetActive(false);
-            }
 
             if (!tableModelHasRails || !hasRunSyncOnce)
-            {
                 PlaceSunkBallsIntoRestingPlace();
-            }
 
             if (logger)
-            {
                 logger._Log(name, "disabling marker because the game is over");
-            }
 
             isRepositioningCueBall = false;
 
             if (marker)
-            {
                 marker.SetActive(false);
-            }
 
             // Remove any access rights
             localPlayerID = -1;
@@ -2163,9 +1453,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             poolMenu._UpdateMainMenuView(
                 isTeams,
                 isTeam2Turn,
-                (int) gameMode,
-                isKorean,
-                isThreeCushionCarom,
+                gameMode,
                 timerSecondsPerShot,
                 player1ID,
                 player2ID,
@@ -2187,17 +1475,18 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         private void OnRemoteTurnChange()
         {
             if (logger)
-                logger._Log(name, "OnRemoteTurnChange");
+                logger._Log(name, nameof(OnRemoteTurnChange));
 
             // Effects
             ApplyTableColour(isTeam2Turn);
             mainSrc.PlayOneShot(newTurnSfx, 1.0f);
             hasFoulBeenPlayedThisTurn = false;
+            isCueOutOfBounds = false;
 
             // Register correct cuetip
             cueTip = cueTips[Convert.ToUInt32(isTeam2Turn)];
 
-            if (isFourBall) // 4 ball
+            if (IsCarom) // 4 ball
             {
                 if (!isTeam2Turn)
                 {
@@ -2248,12 +1537,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
                     // Cope with a marker that might not be synced
                     VRCObjectSync comp = marker.GetComponent<VRCObjectSync>();
                     if (Utilities.IsValid(comp) && Networking.IsOwner(gameObject))
-                    {
                         Networking.SetOwner(Networking.LocalPlayer, marker);
-                        // Move the marker to where the cue ball is, and then teleport it there. This tells VRC to
-                        // move the marker remotely, avoiding weird network sync problems.
-                        // comp.TeleportTo(marker.transform);
-                    }
                 }
             }
 
@@ -2268,9 +1552,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         private void GrantCueAccess()
         {
             if (logger)
-            {
                 logger._Log(name, "GrantCueAccess");
-            }
 
             if (localPlayerID > -1)
             {
@@ -2301,15 +1583,9 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         private void OnRemoteNewGame()
         {
             if (logger)
-            {
                 logger._Log(name, "OnRemoteNewGame");
-            }
 
             poolMenu._EnableResetButton();
-
-            is8Ball = gameMode == 0u;
-            isNineBall = gameMode == 1u;
-            isFourBall = gameMode == 2u;
 
             // Cue ball
             currentBallPositions[0] = new Vector3(-SPOT_POSITION_X, 0.0f, 0.0f);
@@ -2317,221 +1593,41 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
 
             // Start at spot
 
-            if (isNineBall) // 9 ball
-            {
-                Initialize9Ball();
-            }
-            else if (isFourBall) // 4 ball
-            {
-                Initialize4Ball();
-            }
-            else // Normal 8 ball modes
-            {
-                Initialize8Ball();
+            switch (gameMode) {
+                case GameMode.EightBall:
+                    Initialize8Ball();
+                    InitializeEightBallVisuals();
+
+                    break;
+                case GameMode.NineBall:
+                    Initialize9Ball();
+                    InitializeNineBallVisuals();
+
+                    break;
+                case GameMode.KoreanCarom:
+                case GameMode.JapaneseCarom:
+                    Initialize4Ball();
+
+                    InitializeCaromVisuals();
+
+                    break;
+                case GameMode.ThreeCushionCarom:
+                    InitializeThreeCushionCarom();
+                    InitializeThreeCushionCaromVisuals();
+                    
+                    break;
             }
 
             if (localPlayerID >= 0)
-            {
                 playerIsTeam2 = localPlayerID % 2 == 1;
-            }
-
-            if (isNineBall) // 9 Ball / USA colours
-            {
-                pointerColour0 = tableLightBlue;
-                pointerColour1 = tableLightBlue;
-                pointerColour2 = tableLightBlue;
-
-                pointerColourErr = tableBlack; // No error effect
-                pointerClothColour = fabricBlue;
-
-                // 9 ball only uses one colourset / cloth colour
-
-                foreach (MeshRenderer meshRenderer in ballRenderers)
-                {
-                    meshRenderer.material.SetTexture(MainTex, sets[3]);
-                }
-            }
-            else if (isFourBall)
-            {
-                pointerColour0 = tableWhite;
-                pointerColour1 = tableYellow;
-
-                // Should not be used
-                pointerColour2 = tableRed;
-                pointerColourErr = tableRed;
-
-                foreach (MeshRenderer meshRenderer in ballRenderers)
-                {
-                    meshRenderer.material.SetTexture(MainTex, sets[2]);
-                }
-
-                pointerClothColour = fabricGreen;
-            }
-            else // Standard 8 ball derivatives
-            {
-                pointerColourErr = tableRed;
-                pointerColour2 = tableWhite;
-
-                pointerColour0 = tableBlue;
-                pointerColour1 = tableOrange;
-
-                foreach (MeshRenderer meshRenderer in ballRenderers)
-                {
-                    meshRenderer.material.SetTexture(MainTex, sets[0]);
-                }
-
-                pointerClothColour = fabricGray;
-            }
 
             tableRenderer.material.SetColor(ClothColour, pointerClothColour);
 
             if (tableReflection)
-            {
                 tableReflection.RenderProbe();
-            }
 
             ApplyTableColour(false);
             GrantCueAccess();
-
-            if (isNineBall) // 9 ball specific
-            {
-                if (marker9ball)
-                {
-                    marker9ball.SetActive(true);
-                }
-            }
-            else
-            {
-                if (marker9ball)
-                {
-                    marker9ball.SetActive(false);
-                }
-            }
-
-            if (isFourBall) // 4 ball specific
-            {
-                if (pocketBlockers)
-                {
-                    pocketBlockers.SetActive(true);
-                }
-
-                scores[0] = 0;
-                scores[1] = 0;
-
-                // Reset mesh filters on balls that change them
-                ballTransforms[0].GetComponent<MeshFilter>().sharedMesh = cueballMeshes[0];
-                ballTransforms[9].GetComponent<MeshFilter>().sharedMesh = cueballMeshes[1];
-            }
-            else
-            {
-                if (pocketBlockers)
-                {
-                    pocketBlockers.SetActive(false);
-                }
-
-                // Reset mesh filters on balls that change them
-                ballTransforms[0].GetComponent<MeshFilter>().sharedMesh = cueballMeshes[0];
-                ballTransforms[9].GetComponent<MeshFilter>().sharedMesh = nineBall;
-            }
-
-            if (isNineBall)
-            {
-                for (int i = 0; i <= 9; i++)
-                {
-                    if (ballTransforms[i])
-                    {
-                        ballTransforms[i].gameObject.SetActive(true);
-                    }
-
-                    if (ballShadowPosConstraints[i])
-                    {
-                        ballShadowPosConstraints[i].gameObject.SetActive(true);
-                    }
-                }
-
-                for (int i = 10; i < NUMBER_OF_SIMULATED_BALLS; i++)
-                {
-                    if (ballTransforms[i])
-                    {
-                        ballTransforms[i].gameObject.SetActive(false);
-                    }
-
-                    if (ballShadowPosConstraints[i])
-                    {
-                        ballShadowPosConstraints[i].gameObject.SetActive(false);
-                    }
-                }
-            }
-            else if (isFourBall)
-            {
-                for (int i = 1; i < NUMBER_OF_SIMULATED_BALLS; i++)
-                {
-                    if (ballTransforms[i])
-                    {
-                        ballTransforms[i].gameObject.SetActive(false);
-                    }
-
-                    if (ballShadowPosConstraints[i])
-                    {
-                        ballShadowPosConstraints[i].gameObject.SetActive(false);
-                    }
-                }
-
-                if (ballTransforms[0])
-                {
-                    ballTransforms[0].gameObject.SetActive(true);
-                }
-
-                if (ballTransforms[2])
-                {
-                    ballTransforms[2].gameObject.SetActive(true);
-                }
-
-                if (ballTransforms[3])
-                {
-                    ballTransforms[3].gameObject.SetActive(true);
-                }
-
-                if (ballTransforms[9])
-                {
-                    ballTransforms[9].gameObject.SetActive(true);
-                }
-
-                if (ballShadowPosConstraints[0])
-                {
-                    ballShadowPosConstraints[0].gameObject.SetActive(true);
-                }
-
-                if (ballShadowPosConstraints[2])
-                {
-                    ballShadowPosConstraints[2].gameObject.SetActive(true);
-                }
-
-                if (ballShadowPosConstraints[3])
-                {
-                    ballShadowPosConstraints[3].gameObject.SetActive(true);
-                }
-
-                if (ballShadowPosConstraints[9])
-                {
-                    ballShadowPosConstraints[9].gameObject.SetActive(true);
-                }
-            }
-            else
-            {
-                for (int i = 0; i < NUMBER_OF_SIMULATED_BALLS; i++)
-                {
-                    if (ballTransforms[i])
-                    {
-                        ballTransforms[i].gameObject.SetActive(true);
-                    }
-
-                    if (ballShadowPosConstraints[i])
-                    {
-                        ballShadowPosConstraints[i].gameObject.SetActive(true);
-                    }
-                }
-            }
 
             // Effects - don't run this if this is the first network sync, as we may be catching up.
             if (hasRunSyncOnce)
@@ -2551,13 +1647,10 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
 
             // Make sure that we run a pass on rigidbodies to ensure they are off.
             PlaceSunkBallsIntoRestingPlace();
-
             OnRemoteTurnChange();
 
             if (slogger)
-            {
                 slogger.OscReportGameStarted(localPlayerID >= 0);
-            }
         }
 
         /// <summary>
@@ -2568,9 +1661,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         private void PlaceSunkBallsIntoRestingPlace()
         {
             if (logger)
-            {
                 logger._Log(name, "PlaceSunkBallsIntoRestingPlace");
-            }
 
             var numberOfSunkBalls = 0;
             Vector3 localPosition = sunkBallsPositionRoot.localPosition;
@@ -2582,11 +1673,11 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             {
                 ballTransforms[i].GetComponent<Rigidbody>().isKinematic = true;
 
-                if (ballsArePocketed[i])
-                {
-                    ballTransforms[i].localPosition = new Vector3(posX + numberOfSunkBalls * ballDiameter, posY, posZ);
-                    numberOfSunkBalls++;
-                }
+                if (!ballsArePocketed[i]) 
+                    continue;
+                
+                ballTransforms[i].localPosition = new Vector3(posX + numberOfSunkBalls * ballDiameter, posY, posZ);
+                numberOfSunkBalls++;
             }
         }
 
@@ -2596,21 +1687,13 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             for (int i = 2; i < 9; i++)
             {
                 if (!balls[i])
-                {
                     return i;
-                }
             }
 
             if (!balls[1])
-            {
                 return 1;
-            }
-            else if (!balls[9])
-            {
-                return 9;
-            }
 
-            return 0;
+            return !balls[9] ? 9 : 0;
         }
 
         private void OnTurnOverGameWon(bool newIsTeam2Winner, bool causedByFoul)
@@ -2632,9 +1715,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         private void OnTurnOverFoul()
         {
             if (logger)
-            {
                 logger._Log(name, "OnTurnOverFoul");
-            }
 
             isFoul = true;
             isPlayerAllowedToPlay = true;
@@ -2662,23 +1743,15 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         public void _ToggleDesktopUITopDownView()
         {
             if (isInDesktopTopDownView)
-            {
                 OnDesktopTopDownViewExit();
-            }
             else if (canEnterDesktopTopDownView)
-            {
                 OnDesktopTopDownViewStart();
-            }
-
-            Debug.Log("amogus");
         }
 
         private void OnDesktopTopDownViewExit()
         {
             if (logger)
-            {
                 logger._Log(name, "OnDesktopTopDownViewExit");
-            }
 
             isInDesktopTopDownView = false;
 
@@ -2686,7 +1759,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             {
                 desktopBase.SetActive(false);
 
-                if (desktopCamera == null)
+                if (!desktopCamera)
                 {
                     desktopCamera = desktopBase.GetComponentInChildren<Camera>();
                 }
@@ -2711,41 +1784,29 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             pancakeUI._ExitDesktopTopDownView();
         }
 
-        protected abstract void HitBallWithCue(Vector3 shotDirection, float velocity);
-
-        protected void HandleCueBallHit()
+        private void HandleCueBallHit()
         {
             if (logger)
-            {
-                logger._Log(name, "disabling marker because the ball hase been hit");
-            }
+                logger._Log(name, nameof(HandleCueBallHit));
 
             isRepositioningCueBall = false;
 
             if (marker)
-            {
                 marker.SetActive(false);
-            }
 
             if (devhit)
-            {
                 devhit.SetActive(false);
-            }
 
             if (guideline)
-            {
                 guideline.gameObject.SetActive(false);
-            }
 
             // Remove locks
             _EndHit();
             isPlayerAllowedToPlay = false;
             isFoul = false; // In case did not drop foul marker
 
-            float vol = Mathf.Clamp(currentBallVelocities[0].magnitude * 0.1f, 0f, 1f);
-
             cueTipSrc.transform.SetPositionAndRotation(cueTip.transform.position, new Quaternion());
-            cueTipSrc.PlayOneShot(hitBallSfx, vol);
+            cueTipSrc.PlayOneShot(hitBallSfx, Mathf.Clamp(currentBallVelocities[0].magnitude * 0.1f, 0f, 1f));
 
             // Commit changes
             turnIsRunning = true;
@@ -2758,27 +1819,12 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             isSimulatedByUs = true;
         }
 
-        // TODO: Yes this should be at the top of the file but I've not worked on this code for 24 months and am lazy.
-        private Vector3 desktopCameraInitialPosition;
-        private Quaternion desktopCameraInitialRotation;
-        private float lastLookHorizontal;
-        private float lastLookVertical;
-        private bool lastInputUseDown;
-        private float inputHeldDownTime;
-        private float desktopShootForce;
-        private const float desktopShotPowerMult = 0.15f;
-        public GameObject powerBar;
-        public GameObject topBar;
-        public Vector3 initialPowerBarPos;
-
         public override void InputUse(bool value, VRC.Udon.Common.UdonInputEventArgs args)
         {
             if (!isInDesktopTopDownView)
             {
                 if (canEnterDesktopTopDownView)
-                {
                     OnDesktopTopDownViewStart();
-                }
 
                 return;
             }
@@ -2789,9 +1835,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         public override void InputLookHorizontal(float value, VRC.Udon.Common.UdonInputEventArgs args)
         {
             if (!isInDesktopTopDownView)
-            {
                 return;
-            }
 
             lastLookHorizontal = value;
         }
@@ -2799,9 +1843,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         public override void InputLookVertical(float value, VRC.Udon.Common.UdonInputEventArgs args)
         {
             if (!isInDesktopTopDownView)
-            {
                 return;
-            }
 
             lastLookVertical = value;
         }
@@ -2854,10 +1896,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             desktopAimPoint = dir + currentBallPositions[0];
 
             if (!isDesktopLocalTurn)
-            {
-                desktopCamera.transform.SetLocalPositionAndRotation(desktopCameraInitialPosition,
-                    desktopCameraInitialRotation);
-            }
+                desktopCamera.transform.SetLocalPositionAndRotation(desktopCameraInitialPosition, desktopCameraInitialRotation);
             else
             {
                 Vector3 ncursor = desktopAimPoint;
@@ -2885,10 +1924,6 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
 
                         // Create copy of cursor for later
                         desktopSafeRemovePoint = desktopAimPoint;
-
-                        // Unlock cursor position from table
-                        desktopClampX = Mathf.Infinity;
-                        desktopClampY = Mathf.Infinity;
                     }
 
                     // Calculate shoot amount;
@@ -2915,8 +1950,6 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
 
                         // Restore cursor position
                         desktopAimPoint = desktopSafeRemovePoint;
-                        desktopClampX = tableWidth;
-                        desktopClampY = tableHeight;
 
                         // 1-frame override to fix rotation
                         delta = desktopShootVector;
@@ -2954,133 +1987,35 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
                 desktopCamera.transform.localRotation = cue.transform.localRotation;
             }
 
-            Vector3 powerBarPos = powerBar.transform.localPosition;
-            powerBarPos.x = Mathf.Lerp(initialPowerBarPos.x, topBar.transform.localPosition.x,
-                desktopShootForce * 2.0f);
+            var powerBarPos = powerBar.transform.localPosition;
+            powerBarPos.x = Mathf.Lerp(initialPowerBarPos.x, topBar.transform.localPosition.x, desktopShootForce * 2.0f);
             powerBar.transform.localPosition = powerBarPos;
         }
 
         private void UpdateScores()
         {
-            if (isFourBall)
-            {
-                ReportFourBallScore();
+            switch (gameMode) {
+                case GameMode.EightBall:
+                    ReportEightBallScore();
 
-                return;
-            }
+                    return;
+                case GameMode.NineBall:
+                    ReportNineBallScore();
 
-            if (isNineBall)
-            {
-                ReportNineBallScore();
+                    return;
+                case GameMode.KoreanCarom:
+                case GameMode.JapaneseCarom:
+                case GameMode.ThreeCushionCarom:
+                    ReportFourBallScore();
 
-                return;
-            }
-
-            ReportEightBallScore();
-        }
-
-        private void ReportEightBallScore()
-        {
-            if (logger)
-            {
-                logger._Log(name, "ReportEightBallScore");
-            }
-
-            int sunkBlues = 0;
-            int sunkOranges = 0;
-
-            for (int i = 2; i < 9; i++)
-            {
-                if (ballsArePocketed[i])
-                {
-                    sunkBlues++;
-                }
-            }
-
-            for (int i = 9; i < 16; i++)
-            {
-                if (ballsArePocketed[i])
-                {
-                    sunkOranges++;
-                }
-            }
-
-            if (isGameInMenus)
-            {
-                if (isTeam2Winner && ballsArePocketed[1])
-                {
-                    if (isTeam2Blue)
-                    {
-                        sunkBlues++;
-                    }
-                    else
-                    {
-                        sunkOranges++;
-                    }
-                }
-            }
-
-            var teamAScore = isTeam2Blue ? sunkOranges : sunkBlues;
-            var teamBScore = isTeam2Blue ? sunkBlues : sunkOranges;
-
-            poolMenu._SetScore(false, teamAScore);
-            poolMenu._SetScore(true, teamBScore);
-
-            if (slogger)
-            {
-                slogger.OscReportScoresUpdated(isGameOver, turnID, teamAScore, isFoul, teamBScore);
-            }
-        }
-
-        private void ReportNineBallScore()
-        {
-            if (logger)
-            {
-                logger._Log(name, "ReportNineBallScore");
-            }
-
-            poolMenu._SetScore(false, -1);
-            poolMenu._SetScore(true, -1);
-
-            if (slogger)
-            {
-                bool gameOver = false;
-
-                if (isGameInMenus)
-                {
-                    if (ballsArePocketed[1])
-                    {
-                        gameOver = true;
-                    }
-                }
-
-                slogger.OscReportScoresUpdated(gameOver, turnID, -1, isFoul, -1);
-            }
-        }
-
-        private void ReportFourBallScore()
-        {
-            if (logger)
-            {
-                logger._Log(name, "ReportFourBallScore");
-            }
-
-            poolMenu._SetScore(false, scores[0]);
-            poolMenu._SetScore(true, scores[1]);
-
-            if (slogger)
-            {
-                slogger.OscReportScoresUpdated(scores[0] >= 10 || scores[1] >= 10, turnID, scores[0], isFoul,
-                    scores[1]);
+                    return;
             }
         }
 
         private void ResetScores()
         {
             if (logger)
-            {
                 logger._Log(name, "ResetScores");
-            }
 
             poolMenu._SetScore(false, 0);
             poolMenu._SetScore(true, 0);
@@ -3089,26 +2024,17 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         public override void OnPlayerLeft(VRCPlayerApi player)
         {
             if (!Networking.IsOwner(localPlayer, gameObject) || !player.IsValid())
-            {
                 return;
-            }
 
-            int playerID = player.playerId;
+            var playerID = player.playerId;
 
-            if (playerID != player1ID && playerID != player2ID && playerID != player3ID &&
-                playerID != player4ID)
-            {
+            if (playerID != player1ID && playerID != player2ID && playerID != player3ID && playerID != player4ID)
                 return;
-            }
 
             if (isGameInMenus)
-            {
                 RemovePlayerFromGame(playerID);
-            }
             else
-            {
                 _Reset(ResetReason.PlayerLeft);
-            }
         }
 
         /// <Summary> Is the local player near the table? </Summary>
@@ -3150,24 +2076,18 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
         {
             numberOfCuesHeldByLocalPlayer--;
             if (numberOfCuesHeldByLocalPlayer == 0)
-            {
                 CannotEnterDesktopTopDownView();
-            }
         }
 
         private void CheckIfCanEnterDesktopTopDownView()
         {
             if (!isNearTable || numberOfCuesHeldByLocalPlayer <= 0)
-            {
                 return;
-            }
 
             canEnterDesktopTopDownView = true;
 
             if (pancakeUI)
-            {
                 pancakeUI.gameObject.SetActive(true);
-            }
         }
 
         private void CannotEnterDesktopTopDownView()
@@ -3175,9 +2095,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             canEnterDesktopTopDownView = false;
 
             if (pancakeUI)
-            {
                 pancakeUI.gameObject.SetActive(false);
-            }
         }
 
         public static string ToReasonString(ResetReason reason)
@@ -3197,7 +2115,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             }
         }
 
-        protected void HandleBallSunk(bool isSuccess)
+        private void HandleBallSunk(bool isSuccess)
         {
             mainSrc.PlayOneShot(sinkSfx, 1.0f);
 
@@ -3208,7 +2126,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
                 HandleFoulEffects();
         }
 
-        protected void HandleSuccessEffects()
+        private void HandleSuccessEffects()
         {
             // Make a bright flash
             tableCurrentColour *= 1.9f;
@@ -3217,7 +2135,7 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
 
         private bool hasFoulBeenPlayedThisTurn;
 
-        protected void HandleFoulEffects()
+        private void HandleFoulEffects()
         {
             if (hasFoulBeenPlayedThisTurn)
                 return;
@@ -3228,178 +2146,56 @@ namespace VRCBilliardsCE.Packages.com.vrcbilliards.vrcbce.Runtime.Scripts
             PlayAudioClip(foulSfx);
         }
 
-        protected void HandleBallCollision(int ballID, int otherBallID, Vector3 reflection)
+        private void HandleBallCollision(int ballID, int otherBallID, Vector3 reflection)
         {
+            if (logger)
+                logger._Log(name, nameof(HandleBallCollision));
+            
             // Prevent sound spam if it happens
             if (currentBallVelocities[ballID].sqrMagnitude > 0 && currentBallVelocities[otherBallID].sqrMagnitude > 0)
             {
-                int clip = UnityEngine.Random.Range(0, hitsSfx.Length - 1);
-                float vol = Mathf.Clamp01(currentBallVelocities[ballID].magnitude * reflection.magnitude);
                 ballPoolTransforms[ballID].position = ballTransforms[ballID].position;
-                ballPool[ballID].PlayOneShot(hitsSfx[clip], vol);
+                ballPool[ballID].PlayOneShot(hitsSfx[UnityEngine.Random.Range(0, hitsSfx.Length - 1)], Mathf.Clamp01(currentBallVelocities[ballID].magnitude * reflection.magnitude));
             }
 
             if (ballID != 0)
-            {
                 return;
-            }
 
-            if (gameMode == 2u)
-            {
-                if (isKorean)
-                {
+            switch (gameMode) {
+                case GameMode.EightBall:
+                    if (firstHitBallThisTurn == 0) {
+                        firstHitBallThisTurn = otherBallID;
+
+                        if (IsFirstEightBallHitFoul(firstHitBallThisTurn, GetNumberOfSunkBlues(), GetNumberOfSunkOranges()))
+                            HandleFoulEffects();
+                    }
+
+                    break;
+                case GameMode.NineBall:
+                    if (firstHitBallThisTurn == 0) {
+                        firstHitBallThisTurn = otherBallID;
+                    }
+
+                    break;
+                case GameMode.KoreanCarom:
                     HandleKorean4BallScoring(otherBallID);
-                }
-                else if (isThreeCushionCarom)
-                {
-                    HandleThreeCushionCarom(otherBallID);
-                }
-                else
-                {
+
+                    break;
+                case GameMode.JapaneseCarom:
                     HandleJapanese4BallScoring(otherBallID);
-                }
-            }
-            else if (firstHitBallThisTurn == 0)
-            {
-                firstHitBallThisTurn = otherBallID;
 
-                if (is8Ball && IsFirstEightBallHitFoul(firstHitBallThisTurn, GetNumberOfSunkBlues(),
-                        GetNumberOfSunkOranges()))
-                    HandleFoulEffects();
+                     break;
+                case GameMode.ThreeCushionCarom:
+                    HandleThreeCushionCaromScoring(otherBallID);
+
+                    break;
             }
         }
-
-        private bool IsFirstEightBallHitFoul(int hitBallID, int numberOfSunkBlues, int numberOfSunkOranges)
-        {
-            if (isOpen)
-                return hitBallID == 1;
-            
-            var isBlue = isTeam2Turn && isTeam2Blue || !isTeam2Turn && !isTeam2Blue;
-            var winCondition = isBlue ? numberOfSunkBlues == 7 && ballsArePocketed[1] : numberOfSunkOranges == 7 && ballsArePocketed[1];
-
-            // Even if you sunk black, if you hit the wrong ball first you still lose.
-            if (winCondition)
-                return hitBallID != 1;
-            
-            // The seven blue balls are stored at idx 2-8.
-            if (isBlue)
-                return hitBallID < 2 || hitBallID > 8;
-
-            // The seven orange balls are stored at idx 9-15.
-            return hitBallID < 9;
-        }
-
-        private void HandleJapanese4BallScoring(int otherBallID)
-        {
-            if (firstHitBallThisTurn == 0)
-            {
-                firstHitBallThisTurn = otherBallID;
-            }
-            else if (secondBallHitThisTurn == 0)
-            {
-                if (otherBallID == firstHitBallThisTurn)
-                {
-                    return;
-                }
-
-                secondBallHitThisTurn = otherBallID;
-
-                Debug.Log(
-                    $"Scoring a point due to hitting balls {firstHitBallThisTurn} and {secondBallHitThisTurn}");
-                OnLocalCaromPoint(ballTransforms[otherBallID]);
-            }
-            else if (thirdBallHitThisTurn == 0)
-            {
-                if (otherBallID == firstHitBallThisTurn || otherBallID == secondBallHitThisTurn)
-                {
-                    return;
-                }
-
-                thirdBallHitThisTurn = otherBallID;
-                Debug.Log(
-                    $"Scoring a point due to hitting balls {firstHitBallThisTurn} and {secondBallHitThisTurn} and {thirdBallHitThisTurn}");
-                OnLocalCaromPoint(ballTransforms[otherBallID]);
-            }
-        }
-
-        private void HandleKorean4BallScoring(int otherBallID)
-        {
-            if (otherBallID == 9)
-            {
-                if (isMadeFoul)
-                {
-                    return;
-                }
-
-                isMadeFoul = true;
-                scores[Convert.ToUInt32(isTeam2Turn)]--;
-
-                if (scores[Convert.ToUInt32(isTeam2Turn)] < 0)
-                {
-                    scores[Convert.ToUInt32(isTeam2Turn)] = 0;
-                }
-
-                SpawnMinusOne(ballTransforms[otherBallID]);
-            }
-            else if (firstHitBallThisTurn == 0)
-            {
-                firstHitBallThisTurn = otherBallID;
-            }
-            else if (otherBallID != firstHitBallThisTurn)
-            {
-                if (secondBallHitThisTurn == 0)
-                {
-                    secondBallHitThisTurn = otherBallID;
-                    OnLocalCaromPoint(ballTransforms[otherBallID]);
-                }
-            }
-        }
-
-        private void HandleThreeCushionCarom(int otherBallID)
-        {
-            // Three cushion billiards is a normal carom game with one exception: you need to hit THREE cushions during
-            // your carom shot (before you hit the second object ball) for a score to count. It is an extremely hard
-            // game.
-
-            if (otherBallID == 9)
-            {
-                if (isMadeFoul)
-                {
-                    return;
-                }
-
-                // Note we do not penalize in this game. This game is already stupidly hard.
-                isMadeFoul = true;
-            }
-            else if (firstHitBallThisTurn == 0)
-            {
-                firstHitBallThisTurn = otherBallID;
-            }
-            else if (otherBallID != firstHitBallThisTurn && secondBallHitThisTurn == 0)
-            {
-                secondBallHitThisTurn = otherBallID;
-
-                if (cushionsHitThisTurn < 3)
-                {
-                    Debug.Log(
-                        $"Not scoring a point despite hitting {firstHitBallThisTurn} and {secondBallHitThisTurn}: only hit {cushionsHitThisTurn} cushions");
-
-                    return;
-                }
-
-                Debug.Log(
-                    $"Scoring a point by hitting {firstHitBallThisTurn} and {secondBallHitThisTurn}: hit {cushionsHitThisTurn} cushions");
-
-                OnLocalCaromPoint(ballTransforms[otherBallID]);
-            }
-        }
-
+        
         private void PlayAudioClip(AudioClip clip)
         {
-            if (clip == null)
-            {
+            if (!clip)
                 return;
-            }
 
             mainSrc.PlayOneShot(clip);
         }
