+++ Packages/com.vrcbilliards.vrcbce/Runtime/Scripts/PoolStateManager.Base.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9133df8329ab1a646b716a9158e88704
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
