+++ Packages/com.vrcbilliards.vrcbce/Runtime/Scripts/PoolStateManager.Carom.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d086d2c0b97a4a50b9b129a19eec32ce
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
