+++ Packages/com.vrcbilliards.vrcbce/Runtime/Scripts/PoolStateManager.Commands.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9f0133442e8ddc54b8e2a32c1b5b4018
+guid: 04514860a2234c7ca3fd4e47508c50e4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
