+++ Packages/com.vrcbilliards.vrcbce/Runtime/Scripts/PoolStateManager.Debug.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 97423334e41b4787a45fc130aa30e1cb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
