+++ Packages/com.vrcbilliards.vrcbce/Runtime/Scripts/PoolStateManager.EightBall.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 7b64101cf3c64359a734023eba88983c
+timeCreated: 1740012722
