+++ Packages/com.vrcbilliards.vrcbce/Runtime/Scripts/PoolStateManager.ThreeCushionCarom.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0307b32a9a4e7c74faff4809fd255579
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
