+++ Packages/com.vrcbilliards.vrcbce/VRCBCE (akalink).prefab
@@ -24166,7 +24166,8 @@ PrefabInstance:
         type: 3}
       propertyPath: serializedProgramAsset
       value: 
-      objectReference: {fileID: 0}
+      objectReference: {fileID: 11400000, guid: 37ea0e466fc99bf42b4810b336b59f81,
+        type: 2}
     - target: {fileID: 3099494927547820344, guid: 10054966988c6da4f98426aae4877cf4,
         type: 3}
       propertyPath: m_Pivot.x
